dashboard component + side-bar styles
